refactor(bijou): extract filter interactive UI into filter-interactive.ts
Move the ~152-line interactive filter state machine from filter.ts into
a dedicated filter-interactive.ts module:

- filter-interactive.ts: owns FilterOption<T>, FilterOptions<T> types,
  defaultMatch() matcher, and the interactiveFilter() function with
  real-time search, keyboard navigation, and item count display
- filter.ts: slim facade with filter() entry point, fallbackFilter(),
  and type re-exports of FilterOption and FilterOptions

Update ansi-lint.test.ts allowlist:
- Add filter-interactive.ts (inherited ANSI escapes)
- Remove filter.ts (no longer contains ANSI escapes)

Zero test modifications beyond allowlist. Zero index.ts modifications.
